+++ src/components/common/labeled.ts
